requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.152.0
google-genai==0.2.2
requests==2.31.0
apscheduler==3.10.4
python-multipart==0.0.6
Pillow==10.1.0

